move docker-compose to test fixtures - cluster mode works, now need to change test DDL
